Fixed commands and added events
Just did some minor tweaks to the commands to make them work properly.
Main change to that was the implementation of BlockState rather than
Block, since apparently Block is dynamic and the stored value changes
wih the actual block in the server. BlockState, however, takes a
snapshot of the block at that moment, allowing me to use the parts that
matter (Type and MaterialData). Also added event handling that will
automatically delete an arena block when left-clicked by a player who
has permission and will also make sure that a player is allowed to break
an arena block if they try (allowing for a way to protect areas, but
that will be easier when I later implement arena based permissions).
Still have to implement saving arenas, which is why I did not merge this
to master.